Moved class 'HttpClientHandlerBuilder' to 'Builders' folder.
